Moving config under pkg folder
