+++ api/rpc/function/function.go
@@ -1,8 +1,8 @@
 package function
 
 import (
-	"github.com/appcelerator/amp/config"
 	"github.com/appcelerator/amp/data/storage"
+	"github.com/appcelerator/amp/pkg/config"
 	"github.com/appcelerator/amp/pkg/nats-streaming"
 	"github.com/docker/docker/pkg/stringid"
 	"github.com/golang/protobuf/proto"
