+++ api/server/server.go
@@ -19,9 +19,9 @@ import (
 	"github.com/appcelerator/amp/api/rpc/topic"
 	"github.com/appcelerator/amp/api/rpc/version"
 	"github.com/appcelerator/amp/api/runtime"
-	"github.com/appcelerator/amp/config"
 	"github.com/appcelerator/amp/data/influx"
 	"github.com/appcelerator/amp/data/storage/etcd"
+	"github.com/appcelerator/amp/pkg/config"
 	"github.com/docker/docker/client"
 	"google.golang.org/grpc"
 	"os"
