+++ cmd/adm-server/servercore/infrastructure.go
@@ -1,7 +1,7 @@
 package servercore
 
 import (
-	"github.com/appcelerator/amp/config"
+	"github.com/appcelerator/amp/pkg/config"
 	"github.com/docker/docker/api/types/mount"
 	"github.com/docker/docker/api/types/swarm"
 )
