+++ cmd/ampadm/cluster.go
@@ -2,7 +2,7 @@ package main
 
 import (
 	"fmt"
-	"github.com/appcelerator/amp/config"
+	"github.com/appcelerator/amp/pkg/config"
 	"github.com/docker/docker/api/types/mount"
 	"github.com/docker/docker/api/types/swarm"
 )
