+++ tests/integration/data/influx/influx_test.go
@@ -1,8 +1,8 @@
 package influx
 
 import (
-	"github.com/appcelerator/amp/config"
 	. "github.com/appcelerator/amp/data/influx"
+	"github.com/appcelerator/amp/pkg/config"
 	"os"
 	"testing"
 	"time"
