+++ tests/integration/rpc/logs_test.go
@@ -7,7 +7,7 @@ import (
 	"time"
 
 	. "github.com/appcelerator/amp/api/rpc/logs"
-	"github.com/appcelerator/amp/config"
+	"github.com/appcelerator/amp/pkg/config"
 	"github.com/docker/docker/pkg/stringid"
 	"github.com/golang/protobuf/proto"
 	"github.com/nats-io/go-nats"
